# apps/web/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/rooms": {
      "post": {
        "operationId": "CollabController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateCollabRoomDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RoomEntity" } }
            }
          }
        },
        "tags": ["collab"]
      },
      "get": {
        "operationId": "CollabController_findByOwner",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/RoomEntity" }
                }
              }
            }
          }
        },
        "tags": ["collab"]
      }
    },
    "/api/rooms/{slug}": {
      "delete": {
        "operationId": "CollabController_deleteRoom",
        "parameters": [
          { "name": "slug", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["collab"]
      },
      "get": {
        "operationId": "CollabController_findBySlug",
        "parameters": [
          { "name": "slug", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RoomEntity" } }
            }
          }
        },
        "tags": ["collab"]
      },
      "patch": {
        "operationId": "CollabController_updateRoom",
        "parameters": [
          { "name": "slug", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateRoomDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RoomEntity" } }
            }
          }
        },
        "tags": ["collab"]
      }
    },
    "/api/rooms/{slug}/join": {
      "post": {
        "operationId": "CollabController_joinRoom",
        "parameters": [
          { "name": "slug", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/JoinRoomBodyDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JoinRoomResponseDto" }
              }
            }
          }
        },
        "tags": ["collab"]
      }
    },
...
  },
  "info": {
    "title": "Unishare API",
    "description": "API for the Unishare platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
...
      "CreateCollabRoomDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "description": "Optional room title", "maxLength": 120 },
          "visibility": {
            "type": "string",
            "enum": ["OPEN", "VIEW_ONLY", "PRIVATE"],
            "description": "Room visibility (default: OPEN)"
          },
          "password": {
            "type": "string",
            "description": "Optional password to protect room on creation"
          }
        }
      },
      "RoomEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "slug": { "type": "string" },
          "title": { "type": "object", "nullable": true },
          "ownerId": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "visibility": { "type": "string", "enum": ["OPEN", "VIEW_ONLY", "PRIVATE"] },
          "hasPassword": {
            "type": "boolean",
            "description": "Whether room has a password set (never exposes the hash)"
          }
        },
        "required": ["id", "slug", "ownerId", "createdAt", "updatedAt", "visibility", "hasPassword"]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "visibility": { "type": "string", "enum": ["OPEN", "VIEW_ONLY", "PRIVATE"] },
          "title": { "type": "string" },
          "password": {
            "type": "string",
            "nullable": true,
            "description": "Set/change password (string) or remove (null)"
          }
        }
      },
      "JoinRoomBodyDto": {
        "type": "object",
        "properties": {
          "password": { "type": "string" },
          "pwVerified": {
            "type": "boolean",
            "description": "Sentinel from sessionStorage — skip password check if true (D-22)"
          }
        }
      },
      "JoinRoomResponseDto": {
        "type": "object",
        "properties": {
          "roomSlug": { "type": "string" },
          "sessionId": { "type": "string" },
          "userId": { "type": "string" },
          "displayName": { "type": "string" },
          "isAnonymous": { "type": "boolean" },
          "isViewOnly": { "type": "boolean" },
          "ownerId": { "type": "string" }
        },
        "required": [
          "roomSlug",
          "sessionId",
          "userId",
          "displayName",
          "isAnonymous",
          "isViewOnly",
          "ownerId"
        ]
      },
...
    }
  }
}